+++ NEWS.md
@@ -19,6 +19,7 @@
 - Fixed bug in `predict.glmstarma()` that prevented predictions for models with homogeneous intercept.
 - Fixed bug in `glmstarma_predict` that always caused warnings, when time-varying covariates are included in the model
 - Fixed bug in `predict.dglmstarma()` that ignored specified copulas for prediction type `"sample"`
+- Fixed bug in `dglmstarma.control()`, where `parameter_init_dispersion` was not checked
 
 
 # glmsSTARMA 1.0.1